fix: restore endorsed provider cover images on homepage
Move header kind constants out of the client module so RSC serialization keeps kind when EndorsedProviders passes remote covers.
